:recycle: refactor(#168/mandalart): mandalart 도메인 관련 클래스 ninedot-domain 모듈로
