+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/mandalart/exception/HistoryException.java
@@ -1,7 +1,7 @@
 package org.sopt36.ninedotserver.mandalart.exception;
 
-import org.sopt36.ninedotserver.global.exception.BusinessException;
-import org.sopt36.ninedotserver.global.exception.ErrorCode;
+import org.sopt36.ninedotserver.exception.BusinessException;
+import org.sopt36.ninedotserver.exception.ErrorCode;
 
 public class HistoryException extends BusinessException {
