+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/mandalart/model/Cycle.java
@@ -1,4 +1,4 @@
-package org.sopt36.ninedotserver.mandalart.domain;
+package org.sopt36.ninedotserver.mandalart.model;
 
 import lombok.Getter;
